Migrated notifications back to config
