+++ packages/config-plugins/package.json
@@ -34,7 +34,6 @@
   "dependencies": {
     "@expo/config-types": "^40.0.0-beta.2",
     "@expo/configure-splash-screen": "0.4.0",
-    "@expo/image-utils": "0.3.14",
     "@expo/json-file": "8.2.30",
     "@expo/plist": "0.0.13",
     "find-up": "~5.0.0",
