+++ packages/config-plugins/src/android/index.ts
@@ -10,7 +10,6 @@ import * as IntentFilters from './IntentFilters';
 import * as Manifest from './Manifest';
 import * as Name from './Name';
 import * as NavigationBar from './NavigationBar';
-import * as Notifications from './Notifications';
 import * as Orientation from './Orientation';
 import * as Package from './Package';
 import * as Paths from './Paths';
@@ -41,7 +40,6 @@ export {
   IntentFilters,
   Name,
   NavigationBar,
-  Notifications,
   Orientation,
   Package,
   Paths,
